Fix selected instrument indexing in bank export

+++ BambooTracker/bamboo_tracker.cpp
@@ -133,7 +133,7 @@ void BambooTracker::importInstrument(const AbstractBank &bank, size_t index, int
 					   instMan_, std::unique_ptr<AbstractInstrument>(inst)));
 }
 
-void BambooTracker::exportInstruments(BinaryContainer& container, std::vector<size_t> instNums)
+void BambooTracker::exportInstruments(BinaryContainer& container, std::vector<int> instNums)
 {
 	BankIO::saveBank(container, instNums, instMan_);
 }
